Minor gcc cast warning
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@653 37977431-3df6-0310-b722-df95706aa16b

+++ rts/ExternalAI/AICallback.cpp
@@ -642,7 +642,7 @@ float3 CAICallback::ClosestBuildSite(const UnitDef* unitdef,float3 pos,float sea
 	CFeature* feature;
 	int allyteam=gs->AllyTeam(team);
 
-	int endr = searchRadius / (SQUARE_SIZE*2);
+	int endr = (int)(searchRadius / (SQUARE_SIZE*2));
 	const vector<SearchOffset>& ofs = GetSearchOffsetTable (endr);
 
 	for(int so=0;so<endr;so++) {
